issue# 6477.  If old deployed directory doesn't exist, issue warning and continue. QL passed. tested manually. This is merged from the branch version# 23321.
svn path=/trunk/; revision=23368


Former-commit-id: 808fb0b736c2c0ad7ead95791a99dbd5df55f4d3

+++ v3/core/kernel/src/main/java/com/sun/enterprise/v3/admin/adapter/AdminConsoleAdapter.java
@@ -278,6 +278,8 @@ public final class AdminConsoleAdapter extends GrizzlyAdapter implements Adapter
     /**
      *
      */
+    //We will try to backup the old bits, if the old directory doesn't exist, 
+    //issue warning, and continue. see issue# 6477
     private boolean prepareRedeploy(){
         try{
             if (! stopAndCleanup()){
@@ -286,6 +288,10 @@ public final class AdminConsoleAdapter extends GrizzlyAdapter implements Adapter
             }
             File parentFile = warFile.getParentFile();
             File currentDeployedDir = new File( parentFile,ADMIN_APP_NAME);
+            if (! currentDeployedDir.exists()){
+                logger.log(Level.WARNING, currentDeployedDir + " does not exist. Will not do backup for this.");
+                return true;
+            }
             File backupDir = new File(parentFile, ADMIN_APP_NAME+".backup");
             if (currentDeployedDir.renameTo(backupDir))
                 return true;
@@ -295,7 +301,7 @@ public final class AdminConsoleAdapter extends GrizzlyAdapter implements Adapter
         }
         logger.log(Level.SEVERE, "Cannot backup previous version of __admingui ");
         setStateMsg(AdapterState.APPLICATION_BACKUP_FALED);
-        return false;
+        return true;
     }
     
     private void restore(){
